+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/appio/notfication_appio_message.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 """AppIO notification templates"""
-
+from plone import api
 from plone.stringinterp.interfaces import IContextWrapper
 from plone.stringinterp.interfaces import IStringInterpolator
 from Products.DCWorkflow.interfaces import IAfterTransitionEvent
 from zope.component import adapter
 from zope.interface import implementer
 
+from redturtle.prenotazioni import _
 from redturtle.prenotazioni.content.prenotazione import IPrenotazione
 from redturtle.prenotazioni.interfaces import IBookingAPPIoMessage
 from redturtle.prenotazioni.interfaces import IBookingReminderEvent
@@ -29,6 +30,13 @@ class PrenotazioneAPPIoMessage:
 @implementer(IBookingAPPIoMessage)
 @adapter(IPrenotazione, IMovedPrenotazione)
 class PrenotazioneMovedAPPIoMessage(PrenotazioneAPPIoMessage):
+    @property
+    def message_history(self) -> str:
+        _(
+            "history_appio_reschedule_sent",
+            default="AppIO message about the booking reschedule was sent",
+        )
+
     @property
     def message(self) -> str:
         return IStringInterpolator(IContextWrapper(self.prenotazione)())(
@@ -54,6 +62,22 @@ class PrenotazioneMovedAPPIoMessage(PrenotazioneAPPIoMessage):
 @adapter(IPrenotazione, IAfterTransitionEvent)
 class PrenotazioneAfterTransitionAPPIoMessage(PrenotazioneAPPIoMessage):
     @property
+    def message_history(self) -> str:
+        transition = (
+            self.event.transition
+            and api.portal.translate(self.event.transition.title)
+            or ""
+        )
+
+        return api.portal.translate(
+            _(
+                "history_appio_transition_sent",
+                "AppIO message about the ${transition} transition was sent.",
+                mapping={"transition": transition},
+            ),
+        )
+
+    @property
     def message(self) -> str:
         return IStringInterpolator(IContextWrapper(self.prenotazione)())(
             getattr(
@@ -77,6 +101,12 @@ class PrenotazioneAfterTransitionAPPIoMessage(PrenotazioneAPPIoMessage):
 @implementer(IBookingAPPIoMessage)
 @adapter(IPrenotazione, IBookingReminderEvent)
 class PrenotazioneReminderAppIOMessage(PrenotazioneAPPIoMessage):
+    @property
+    def message_history(self) -> str:
+        return api.portal.translate(
+            _("history_appio_reminder_sent", default="AppIO reminder was sent")
+        )
+
     @property
     def message(self) -> str:
         return IStringInterpolator(IContextWrapper(self.prenotazione)())(