+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/sms/events.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from functools import partial
 
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
@@ -7,6 +8,7 @@ from redturtle.prenotazioni.interfaces import IBookingNotificationSender
 from redturtle.prenotazioni.interfaces import IBookingSMSMessage
 from redturtle.prenotazioni.utilities import handle_exception_by_log
 
+from .. import notify_the_message_failure
 from . import INotificationSMS
 
 
@@ -14,7 +16,11 @@ def booking_folder_provides_current_behavior(booking):
     return INotificationSMS.providedBy(booking.getPrenotazioniFolder())
 
 
+notify_the_message_failure = partial(notify_the_message_failure, gateway_type="SMS")
+
+
 @handle_exception_by_log
+@notify_the_message_failure
 def send_notification_on_transition(context, event) -> None:
     if not booking_folder_provides_current_behavior(context):
         return
@@ -45,6 +51,7 @@ def send_notification_on_transition(context, event) -> None:
 
 
 @handle_exception_by_log
+@notify_the_message_failure
 def notify_on_move(context, event):
     if not booking_folder_provides_current_behavior(context):
         return
@@ -69,6 +76,7 @@ def notify_on_move(context, event):
 
 
 @handle_exception_by_log
+@notify_the_message_failure
 def send_booking_reminder(context, event):
     if not booking_folder_provides_current_behavior(context):
         return