+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/sms/message.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 """SMS Notification Templates"""
-
+from plone import api
 from plone.stringinterp.interfaces import IContextWrapper
 from plone.stringinterp.interfaces import IStringInterpolator
 from Products.DCWorkflow.interfaces import IAfterTransitionEvent
 from zope.component import adapter
 from zope.interface import implementer
 
+from redturtle.prenotazioni import _
 from redturtle.prenotazioni.content.prenotazione import IPrenotazione
 from redturtle.prenotazioni.interfaces import IBookingReminderEvent
 from redturtle.prenotazioni.interfaces import IBookingSMSMessage
@@ -23,7 +24,11 @@ class PrenotazioneSMSMessage:
 
     @property
     def message(self) -> str:
-        NotImplementedError
+        raise NotImplementedError("The method was not implemented")
+
+    @property
+    def message_history(self) -> str:
+        raise NotImplementedError("The method was not implemented")
 
 
 @implementer(IBookingSMSMessage)
@@ -37,6 +42,13 @@ class PrenotazioneMovedSMSMessage(PrenotazioneSMSMessage):
                 "notify_on_move_sms_message",
                 None,
             ),
+        )
+
+    @property
+    def message_history(self) -> str:
+        _(
+            "history_sms_reschedule_sent",
+            default="SMS message about the booking reschedule was sent",
         )
 
 
@@ -53,6 +65,21 @@ class PrenotazioneAfterTransitionSMSMessage(PrenotazioneSMSMessage):
             ),
         )
 
+    @property
+    def message_history(self) -> str:
+        transition = (
+            self.event.transition
+            and api.portal.translate(self.event.transition.title)
+            or ""
+        )
+        return api.portal.translate(
+            _(
+                "history_sms_transition_sent",
+                "SMS message about the ${transition} transition was sent",
+                mapping={"transition": transition},
+            ),
+        )
+
 
 @implementer(IBookingSMSMessage)
 @adapter(IPrenotazione, IBookingReminderEvent)
@@ -65,4 +92,10 @@ class PrenotazioneReminderSMSMessage(PrenotazioneSMSMessage):
                 "notify_as_reminder_sms_message",
                 "",
             )
+        )
+
+    @property
+    def message_history(self) -> str:
+        return api.portal.translate(
+            _("history_sms_reminder_sent", default="SMS reminder was sent")
         )